fix: Move new enum mappings to new file - and removed the ones from config_legacy that are now only used in the new config system.
