+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/NativeRender.vcxproj
@@ -108,6 +108,7 @@
   <ItemGroup>
     <ClInclude Include="include\Config\ClientCredentials.hpp" />
     <ClInclude Include="include\Config\Config.hpp" />
+    <ClInclude Include="include\Config\EnumMappings.h" />
     <ClInclude Include="include\Config\PlatformConfig.hpp" />
     <ClInclude Include="include\Config\ProductConfig.hpp" />
     <ClInclude Include="include\Config\ProductionEnvironments.hpp" />
